swiper/requirements.txt
amqp==2.5.1
appnope==0.1.0
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==1.11.11
django-crontab==0.7.1
django-redis==4.10.0
fdfs-client-py==1.2.6
gevent==1.4.0
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
kombu==4.6.4
more-itertools==7.2.0
mutagen==1.42.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycryptodomex==3.7.2
Pygments==2.4.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.1
pytz==2019.2
qiniu==7.2.6
redis==3.3.8
requests==2.22.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
zipp==0.6.0
